Rename FastBackend => Indexedbackend, and move some tests into the package they test.
