+++ container-core/src/main/resources/configdefinitions/container.qr-searchers.def
@@ -25,7 +25,7 @@ com.yahoo.prelude.searcher.XMLStringSearcher.source string default=""
 
 ## Default docsum class the QR server should ask the backend to
 ## use for representing hints as default.
-com.yahoo.prelude.fastsearch.FastBackend.docsum.defaultclass string default=""
+com.yahoo.prelude.fastsearch.IndexedBackend.docsum.defaultclass string default=""
 
 com.yahoo.prelude.querytransform.PhrasingSearcher.automatonfile string default=""
 com.yahoo.prelude.querytransform.NonPhrasingSearcher.automatonfile string default=""
