+++ container-search/src/test/java/com/yahoo/prelude/fastsearch/PartialFillTestCase.java
@@ -1,12 +1,8 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.prelude.fastsearch.test;
+package com.yahoo.prelude.fastsearch;
 
-import com.yahoo.component.chain.Chain;
-import com.yahoo.prelude.fastsearch.FastHit;
-import com.yahoo.prelude.fastsearch.VespaBackend;
 import com.yahoo.search.Query;
 import com.yahoo.search.Result;
-import com.yahoo.search.Searcher;
 import com.yahoo.search.result.ErrorHit;
 import com.yahoo.search.result.ErrorMessage;
 import org.junit.jupiter.api.Test;
@@ -139,12 +135,6 @@ public class PartialFillTestCase {
 
     private void doFill(VespaBackend searcher, Result result, String summaryClass) {
         searcher.fill(result, summaryClass);
-    }
-
-    private Chain<Searcher> chainedAsSearchChain(Searcher topOfChain) {
-        List<Searcher> searchers = new ArrayList<>();
-        searchers.add(topOfChain);
-        return new Chain<>(searchers);
     }
 
 }
